requirements.txt
opencv-python-headless
streamlit
mediapipe==0.9.1.0
numpy
Pillow==9.2.0
pandas
tensorflow
matplotlib
tensorflow==2.10.0
fastapi
pytz
uvicorn